+++ plugin/src/main/groovy/org/openbakery/XcodePlugin.groovy
@@ -33,6 +33,7 @@ import org.openbakery.configuration.XcodeConfigTask
 import org.openbakery.coverage.CoverageCleanTask
 import org.openbakery.coverage.CoveragePluginExtension
 import org.openbakery.coverage.CoverageTask
+import org.openbakery.cpd.CpdTask
 import org.openbakery.crashlytics.CrashlyticsPluginExtension
 import org.openbakery.crashlytics.CrashlyticsUploadTask
 import org.openbakery.deploygate.DeployGateCleanTask
@@ -59,6 +60,9 @@ import org.openbakery.signing.ProvisioningInstallTask
 import org.openbakery.simulators.SimulatorsCleanTask
 import org.openbakery.simulators.SimulatorsCreateTask
 import org.openbakery.simulators.SimulatorsListTask
+import org.openbakery.simulators.SimulatorsStartTask
+import org.openbakery.simulators.SimulatorsRunAppTask
+import org.openbakery.simulators.SimulatorsInstallAppTask
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
@@ -84,6 +88,9 @@ class XcodePlugin implements Plugin<Project> {
 	public static final String SIMULATORS_LIST_TASK_NAME = "simulatorsList"
 	public static final String SIMULATORS_CREATE_TASK_NAME = "simulatorsCreate"
 	public static final String SIMULATORS_CLEAN_TASK_NAME = "simulatorsClean"
+	public static final String SIMULATORS_START_TASK_NAME = "simulatorsStart"
+	public static final String SIMULATORS_INSTALL_APP_TASK_NAME = "simulatorsInstallApp"
+	public static final String SIMULATORS_RUN_APP_TASK_NAME = "simulatorsRunApp"
 	public static final String XCODE_BUILD_TASK_NAME = "xcodebuild"
 	public static final String XCODE_CLEAN_TASK_NAME = "xcodebuildClean"
 	public static final String XCODE_CONFIG_TASK_NAME = "xcodebuildConfig"
@@ -111,12 +118,16 @@ class XcodePlugin implements Plugin<Project> {
 	public static final String COCOAPODS_TASK_NAME = 'cocoapods'
 	public static final String OCLINT_TASK_NAME = 'oclint'
 	public static final String OCLINT_REPORT_TASK_NAME = 'oclintReport'
+	public static final String CPD_TASK_NAME = 'cpd'
+	public static final String ENTITLEMENTSMODIFY_TASK_NAME = 'entitlementsModify'
 
 	public static final String APPLEDOC_TASK_NAME = 'appledoc'
 	public static final String APPLEDOC_CLEAN_TASK_NAME = 'appledocClean'
 
 	public static final COVERAGE_TASK_NAME = 'coverage'
 	public static final COVERAGE_CLEAN_TASK_NAME = 'coverageClean'
+
+
 	public static final String SDK_MACOSX = "macosx"
 	public static final String SDK_IPHONEOS = "iphoneos"
 	public static final String SDK_IPHONESIMULATOR = "iphonesimulator"
@@ -137,6 +148,7 @@ class XcodePlugin implements Plugin<Project> {
 		configureHockeyKit(project)
 		configureKeychain(project)
 		configureInfoPlist(project)
+		configureEntitlements(project)
 		configureProvisioning(project)
 		configureAppstore(project)
 		configureHockeyApp(project)
@@ -145,6 +157,7 @@ class XcodePlugin implements Plugin<Project> {
 		configurePackage(project)
 		configureAppledoc(project)
 		configureCoverage(project)
+		configureCpd(project)
 		configureCocoapods(project)
 		configureOCLint(project)
 		configureSimulatorTasks(project)
@@ -265,6 +278,9 @@ class XcodePlugin implements Plugin<Project> {
 			if (project.hasProperty('xcodebuild.ipaFileName')) {
 				project.xcodebuild.ipaFileName = project['xcodebuild.ipaFileName']
 			}
+			if (project.hasProperty('xcodebuild.appExtensions')) {
+				project.xcodebuild.appExtensions = project['xcodebuild.appExtensions']
+			}
 
 			if (project.hasProperty('hockeykit.displayName')) {
 				project.hockeykit.displayName = project['hockeykit.displayName']
@@ -371,6 +387,29 @@ class XcodePlugin implements Plugin<Project> {
 			}
 
 
+			if (project.hasProperty('oclint.reportType')) {
+				project.oclint.reportType = project['oclint.reportType'];
+			}
+			if (project.hasProperty('oclint.rules')) {
+				project.oclint.rules = project['oclint.rules'];
+			}
+			if (project.hasProperty('oclint.disableRules')) {
+				project.oclint.disableRules = project['oclint.disableRules'];
+			}
+			if (project.hasProperty('oclint.excludes')) {
+				project.oclint.excludes = project['oclint.excludes'];
+			}
+			if (project.hasProperty('oclint.maxPriority1')) {
+				project.oclint.maxPriority1 = project['oclint.maxPriority1'];
+			}
+			if (project.hasProperty('oclint.maxPriority2')) {
+				project.oclint.maxPriority2 = project['oclint.maxPriority2'];
+			}
+			if (project.hasProperty('oclint.maxPriority3')) {
+				project.oclint.maxPriority3 = project['oclint.maxPriority3'];
+			}
+
+
 			Task testTask = (Test) project.getTasks().findByPath(JavaPlugin.TEST_TASK_NAME)
 			if (testTask == null) {
 				testTask = project.getTasks().create(JavaPlugin.TEST_TASK_NAME)
@@ -422,6 +461,9 @@ class XcodePlugin implements Plugin<Project> {
 		project.task(SIMULATORS_LIST_TASK_NAME, type: SimulatorsListTask, group: SIMULATORS_LIST_TASK_NAME)
 		project.task(SIMULATORS_CREATE_TASK_NAME, type: SimulatorsCreateTask, group: SIMULATORS_LIST_TASK_NAME)
 		project.task(SIMULATORS_CLEAN_TASK_NAME, type: SimulatorsCleanTask, group: SIMULATORS_LIST_TASK_NAME)
+		project.task(SIMULATORS_START_TASK_NAME, type: SimulatorsStartTask, group: SIMULATORS_LIST_TASK_NAME)
+		project.task(SIMULATORS_RUN_APP_TASK_NAME, type: SimulatorsRunAppTask, group: SIMULATORS_LIST_TASK_NAME)
+		project.task(SIMULATORS_INSTALL_APP_TASK_NAME, type: SimulatorsInstallAppTask, group: SIMULATORS_LIST_TASK_NAME)
 	}
 
 	private void configureHockeyKit(Project project) {
@@ -498,6 +540,9 @@ class XcodePlugin implements Plugin<Project> {
 
 	}
 
+	private void configureCpd(Project project) {
+		project.task(CPD_TASK_NAME, type: CpdTask, group: ANALYTICS_GROUP_NAME)
+	}
 
 	private void configureDeployGate(Project project) {
 		project.task(DEPLOYGATE_CLEAN_TASK_NAME, type: DeployGateCleanTask, group: DEPLOYGATE_GROUP_NAME)
@@ -517,6 +562,7 @@ class XcodePlugin implements Plugin<Project> {
 		}
 	}
 
+
 	private void configureOCLint(Project project) {
 		OCLintTask reportTask = project.task(OCLINT_REPORT_TASK_NAME, type: OCLintTask, group: ANALYTICS_GROUP_NAME)
 
@@ -533,6 +579,10 @@ class XcodePlugin implements Plugin<Project> {
 
 	}
 
+
+	private void configureEntitlements(Project project) {
+		project.task(ENTITLEMENTSMODIFY_TASK_NAME, type: EntitlementsModifyTask, group: XCODE_GROUP_NAME)
+	}
 
 	private void addDependencyToBuild(Project project, Task task) {
 		XcodeBuildTask buildTask = project.getTasks().getByName(XCODE_BUILD_TASK_NAME)