+++ plugin/src/main/groovy/org/openbakery/crashlytics/CrashlyticsUploadTask.groovy
@@ -19,13 +19,14 @@ import org.gradle.api.tasks.TaskAction
 import org.gradle.logging.StyledTextOutput
 import org.gradle.logging.StyledTextOutputFactory
 import org.openbakery.AbstractDistributeTask
+import org.openbakery.XcodePlugin
 import org.openbakery.output.ConsoleOutputAppender
 
 class CrashlyticsUploadTask extends AbstractDistributeTask {
 
 	CrashlyticsUploadTask() {
 		super()
-		dependsOn()
+		dependsOn(XcodePlugin.PACKAGE_TASK_NAME)
 		this.description = "Upload the IPA to crashlytics for crash reports"
 	}
 
@@ -78,8 +79,10 @@ class CrashlyticsUploadTask extends AbstractDistributeTask {
 			commandList.push(project.crashlytics.notifications ? 'YES' : 'NO')
 		}
 
+		def environment = ["DEVELOPER_DIR":project.xcodebuild.xcodePath + "/Contents/Developer/"]
+
 		StyledTextOutput output = getServices().get(StyledTextOutputFactory.class).create(CrashlyticsUploadTask.class)
-		commandRunner.run(commandList, new ConsoleOutputAppender(output))
+		commandRunner.run(commandList, environment, new ConsoleOutputAppender(output))
 
 
 	}