+++ plugin/src/test/groovy/org/openbakery/signing/KeychainCreateTaskTest.groovy
@@ -1,13 +1,11 @@
 package org.openbakery.signing
 
-import org.apache.commons.io.FileUtils
 import org.gmock.GMockController
 import org.gradle.api.Project
 import org.gradle.testfixtures.ProjectBuilder
+import org.junit.AfterClass
 import org.openbakery.CommandRunner
 import org.openbakery.Version
-import org.openbakery.signing.KeychainCreateTask
-import org.testng.annotations.AfterTest
 import org.testng.annotations.BeforeClass
 import org.testng.annotations.Test
 
@@ -42,7 +40,7 @@ class KeychainCreateTaskTest {
 		keychainDestinationFile = new File(project.xcodebuild.signing.signingDestinationRoot, certificateFile.getName())
 	}
 
-	@AfterTest
+	@AfterClass
 	def cleanAfterTest() {
 		certificateFile.delete();
 		new File(project.xcodebuild.signing.signingDestinationRoot, certificateFile.getName()).delete()